refactor/small refactor of user page
