+++ src/components/UserProfile/SavedRecipesSection.tsx
@@ -10,7 +10,7 @@ import {
   RecipeCard,
   type RecipeCardProps,
 } from "@/components/ui-custom/RecipeCard";
-
+// TODO - i left it for now, but this section is probably to delete
 export const SavedRecipesSection = () => {
   const data: RecipeCardProps[] = [
     {
